api/requirements.txt
flask
flask-cors
requests
beautifulsoup4
anthropic
python-dotenv
PyJWT
google-auth
google-auth-oauthlib
